+++ packages/themes/src/themes/shadesOfPurple.ts
@@ -2,6 +2,7 @@ import { experimental_createTheme } from '../createTheme';
 import { dark } from './dark';
 
 export const shadesOfPurple = experimental_createTheme({
+  name: 'shadesOfPurple',
   baseTheme: dark,
   variables: {
     colorBackground: '#3f3c77',